requirements.txt
isodate
celery[redis]
fastapi==0.95.1
google==3.0.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
python-dotenv==0.19.2
redis==4.5.4
requests==2.28.2
requests-oauthlib==1.3.1
uvicorn==0.21.1
redis-server